Fix iframe skeleton, fix react fast refresh warning
